// frontend/src/shared/types.d.ts
type Image = {
  link?: string;
  src: string | StaticImageData;
  alt: string;
};

type Item = {
  title?: string | boolean | number;
  description?: string | Array<string>;
  href?: string;
  form?: SmallForm;
  icon?: Icon;
  callToAction?: CallToActionType;
};
// ...
type Team = {
  name: string;
  occupation: string;
  image: Image;
  items?: Array<Item>;
  containerClass?: string;
  imageClass?: string;
  panelClass?: string;
  nameClass?: string;
  occupationClass?: string;
  itemsClass?: string;
};
// ...
export type TeamHeader = {
  title: string;
  subtitle: string;
};

// Main teamHome export type
export type TeamProps = {
  id: string;
  hasBackground: boolean;
  header: TeamHeader;
  teams: Team[];
};